pip install opencv-python pip install numpy
This project is a Face Recognition-Based Attendance System that uses OpenCV and Haar cascades to detect and recognize faces via a webcam. It stores user details (ID, name) in a CSV file and tracks attendance in date-specific files. Built with Tkinter, it features a user-friendly GUI for registration, attendance, and password management.